+++ src/main/java/name/divinityunbound/datagen/ModLootTableProvider.java
@@ -1,6 +1,7 @@
 package name.divinityunbound.datagen;
 
 import name.divinityunbound.block.ModBlocks;
+import name.divinityunbound.block.custom.EtherealCrystalGreensBlock;
 import name.divinityunbound.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
@@ -9,12 +10,14 @@ import net.minecraft.data.server.loottable.BlockLootTableGenerator;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.item.Item;
 import net.minecraft.loot.LootTable;
+import net.minecraft.loot.condition.BlockStatePropertyLootCondition;
 import net.minecraft.loot.entry.ItemEntry;
 import net.minecraft.loot.entry.LeafEntry;
 import net.minecraft.loot.entry.LootPoolEntry;
 import net.minecraft.loot.function.ApplyBonusLootFunction;
 import net.minecraft.loot.function.SetCountLootFunction;
 import net.minecraft.loot.provider.number.UniformLootNumberProvider;
+import net.minecraft.predicate.StatePredicate;
 
 public class ModLootTableProvider extends FabricBlockLootTableProvider {
     public ModLootTableProvider(FabricDataOutput dataOutput) {
@@ -85,6 +88,12 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.MOB_ATTRACTOR);
 
         addDrop(ModBlocks.SPACE_TIME_FURNACE);
+
+        BlockStatePropertyLootCondition.Builder builder = BlockStatePropertyLootCondition.builder(ModBlocks.ETHEREAL_CRYSTAL_GREENS)
+                .properties(StatePredicate.Builder.create()
+                        .exactMatch(EtherealCrystalGreensBlock.AGE, 3));
+        addDrop(ModBlocks.ETHEREAL_CRYSTAL_GREENS, cropDrops(ModBlocks.ETHEREAL_CRYSTAL_GREENS, ModItems.ETHEREAL_CRYSTAL_GREENS, ModItems.ETHEREAL_CRYSTAL_GREENS_SEEDS, builder));
+
     }
 
     public LootTable.Builder copperLikeOreDrops(Block drop, Item item) {